refactor(config_routes): extract ui.py from _factory.py (6/8)
Schritt 6 von 8 im Split-Plan
(`docs/superpowers/plans/2026-04-29-config-routes-split.md`).

What moved
- `_register_ui_routes()` (~966 LOC, allein wegen Groesse) ->
  `cognithor.channels.config_routes.ui`. Control-Center-Endpoints fuer
  das Flutter Frontend: System-Status, Models-Discovery, Templates,
  Connectors, Wizards-UI, Theme-Settings, Cron-Jobs, MCP-Servers,
  A2A-Config, Budget/Resources, Evolution. Inkl. lokale `_load_yaml` /
  `_save_yaml` Helfer (closures innerhalb der Funktion).

`_factory.py` schrumpft von 2 917 -> 1 952 Zeilen (-33 % in dieser PR).

Per-Sub-Module-Imports (`ui.py`)
- Stdlib: `asyncio`, `time`. `pathlib.Path` nur in TYPE_CHECKING (mit `from __future__ import annotations` sicher).
- Drittpaket: `yaml`.
- Try-Imports: `Request`, `HTTPException`.
- Typing: `Any`, `ConfigManager` (TYPE_CHECKING). F401-clean.

`_factory.py` Cleanup
- `time`, `yaml` Top-Level-Imports entfernt — beide wandern mit `_register_ui_routes` nach `ui.py`. `_factory.py` braucht sie nicht mehr.
- `from pathlib import Path` (war in TYPE_CHECKING) entfernt — die `_load_yaml`/`_save_yaml` Closures, die `Path` als Annotation brauchten, sind jetzt in `ui.py`.

Hinweis: ruff RUF006 fuer `asyncio.create_task(gateway.shutdown())` mit `# noqa: RUF006` unterdrueckt (fire-and-forget; der Shutdown-Task beendet den Prozess, dangling reference ist intentional). Identisch zum Verhalten im Original.

Verification
- `pytest test_route_inventory_unchanged` -> passed (303/303 endpoints).
- Konsumenten-Sweep ueber 11 Test-Files -> 396 passed.
- `ruff check` -> All checks passed.
- `ruff format --check` -> 10 files unchanged.
